Extend standard Spine migrations
The `MarkArchived`, `MarkDeleted` and `RemoveFromStorage` migrations are now exposed as standard Spine migrations for both projections and process managers.

`ProjectionColumnsUpdate` and `ProcessManagerColumnsUpdate` renamed to `UpdateColumns`.
